// projects/08-todo-app-ts/src/types.d.ts
import { TODO_FILTERS } from "./consts";

export interface Todo {
  id: string;
  title: string;
  completed: boolean;
}

export type TodoId = Pick<Todo, "id">;
export type TodoTitle = Pick<Todo, "title">;
export type TodoCompleted = Pick<Todo, "completed">;
export type TodoIdCompleted = Omit<Todo, "title">;

export type ListOfTodos = Todo[]; // para decir que utilizamos types xD

export type FilterValue = (typeof TODO_FILTERS)[keyof typeof TODO_FILTERS];